Swap inheritance for JsonComparison and ConfigurableResponseComparison
Swap class hierachy to make future API clearer.
